rename cggnn submodule to graphs
