+++ pyproject.toml.unversioned
@@ -42,7 +42,7 @@ apiserver = [
     "pydantic==2.0.2",
     "secure==0.3.0"
 ]
-cggnn = [
+graphs = [
     "bokeh==3.3.1",
     "numpy==1.24.3",
     "scipy==1.10.1",
@@ -116,8 +116,8 @@ packages = [
     "spatialprofilingtoolbox.apiserver",
     "spatialprofilingtoolbox.apiserver.app",
     "spatialprofilingtoolbox.apiserver.scripts",
-    "spatialprofilingtoolbox.cggnn",
-    "spatialprofilingtoolbox.cggnn.scripts",
+    "spatialprofilingtoolbox.graphs",
+    "spatialprofilingtoolbox.graphs.scripts",
     "spatialprofilingtoolbox.ondemand",
     "spatialprofilingtoolbox.ondemand.providers",
     "spatialprofilingtoolbox.ondemand.scripts",
@@ -193,7 +193,7 @@ packages = [
     "refresh_views.sql",
     "feature_descriptions.tsv"
 ]
-"spatialprofilingtoolbox.cggnn.scripts" = [
+"spatialprofilingtoolbox.graphs.scripts" = [
     "create_specimen_graphs.py",
     "explore_classes.py",
     "extract.py",
