+++ spatialprofilingtoolbox/db/accessors/__init__.py
@@ -1,6 +1,6 @@
 """Convenience classes for accessing various data in the database."""
 
-from spatialprofilingtoolbox.db.accessors.cggnn import CGGNNAccess
+from spatialprofilingtoolbox.db.accessors.graphs import GraphsAccess
 from spatialprofilingtoolbox.db.accessors.phenotypes import PhenotypesAccess
 from spatialprofilingtoolbox.db.accessors.study import StudyAccess
 from spatialprofilingtoolbox.db.accessors.umap import UMAPAccess
