+++ spatialprofilingtoolbox/graphs/README.md
@@ -1,4 +1,4 @@
-# cggnn
+# graphs
 
 This module constructs cell graphs and uses neural networks on the created graphs to predict outcomes of your choice. It is dependent on the [cg-gnn on pip](https://github.com/CarlinLiao/cg-gnn).
