+++ spatialprofilingtoolbox/graphs/template.config
