+++ test/cggnn/unit_tests/test_explore.sh
@@ -1,8 +1,8 @@
-spt cggnn explore-classes \
+spt graphs explore-classes \
     --database-config-file ../db/.spt_db.config.container \
     --study-name "Melanoma intralesional IL2"
 status=$?
-[ $status -eq 0 ] || echo "cggnn explore-classes failed."
+[ $status -eq 0 ] || echo "graphs explore-classes failed."
 
 if [ $status -eq 0 ];
 then
