+++ test/cggnn/unit_tests/test_extract.sh
@@ -1,9 +1,9 @@
-spt cggnn extract \
+spt graphs extract \
     --config_path unit_tests/spt-test.config \
     --output_directory .
 $([ $? -eq 0 ] && [ -e "label_to_result.json" ] && [ -e "cells.h5" ] && [ -e "labels.h5" ])
 status="$?"
-[ $status -eq 0 ] || echo "cggnn extract failed."
+[ $status -eq 0 ] || echo "graphs extract failed."
 
 cat "label_to_result.json"
 python3.11 -c 'import pandas as pd; print(pd.read_hdf("cells.h5"))'
